requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
easyocr>=1.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.3.0

